+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 ## Fixed
 - Fixed UI Add New Product form
 - Fixed Catalog list for Customer
+- Fixed UI for option/confirm modal